\name{mcmc-package}
\alias{mcmc-package}
\alias{mcmc}
\docType{package}
\title{
\packageTitle{mcmc}
}
\description{
\packageDescription{mcmc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mcmc}
\packageIndices{mcmc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{mcmc}

Maintainer: \packageMaintainer{mcmc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
